django_app_basic1/requirements.txt
asgiref==3.3.4
Django==2.1.7
django-crispy-forms==1.11.2
django-multiselectfield==0.1.12
django-widget-tweaks==1.4.8
html-json-forms==1.1.1
package-name==0.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
